+++ libraries/template-haskell/Language/Haskell/TH/Lib/Internal.hs
@@ -287,10 +287,16 @@ lamCaseE :: [MatchQ] -> ExpQ
 lamCaseE ms = sequence ms >>= return . LamCaseE
 
 tupE :: [ExpQ] -> ExpQ
-tupE es = do { es1 <- sequence es; return (TupE es1)}
+tupE es = do { es1 <- sequence es; return (TupE $ map Just es1)}
+
+tupE' :: [Maybe ExpQ] -> ExpQ
+tupE' es = do { es1 <- traverse sequence es; return (TupE es1)}
 
 unboxedTupE :: [ExpQ] -> ExpQ
-unboxedTupE es = do { es1 <- sequence es; return (UnboxedTupE es1)}
+unboxedTupE es = do { es1 <- sequence es; return (UnboxedTupE $ map Just es1)}
+
+unboxedTupE' :: [Maybe ExpQ] -> ExpQ
+unboxedTupE' es = do { es1 <- traverse sequence es; return (UnboxedTupE es1)}
 
 unboxedSumE :: ExpQ -> SumAlt -> SumArity -> ExpQ
 unboxedSumE e alt arity = do { e1 <- e; return (UnboxedSumE e1 alt arity) }